IS-2573: Add view when no historic meetings
